<==== Mercurial data ====>
Branch: nat_traversal

+++ common/src/utils/media/media_stream_cache.cpp
@@ -377,7 +377,7 @@ void MediaStreamCache::unblockData( int blockingID )
     m_dataBlockings.erase( it );
 }
 
-size_t MediaStreamCache::inactivityPeriod() const
+qint64 MediaStreamCache::inactivityPeriod() const
 {
     QnMutexLocker lk( &m_mutex );
     return m_inactivityTimer.elapsed();
